Move light files to appropriate locations
